+++ src/main/java/com/github/lithualien/data/Asteroid.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.github.lithualien.Classes;
+package com.github.lithualien.data;
 
 
 import java.text.NumberFormat;
