+++ src/main/java/com/github/lithualien/data/Close_approach.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.github.lthualien.Classes;
+package com.github.lithualien.data;
 
 /**
  *
